# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the document for The documentation of My portfolio API",
    "version": "1.0.0",
    "title": "Portfolio API",
    "contact": {
      "email": "inno100os@gmail.com"
    }
  },
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/api/user/blogs/{blogId}": {
      "patch": {
        "tags": ["blog"],
        "summary": "Update an existing blog",
        "description": "",
        "operationId": "updateBlog",
        "consumes": ["application/json"],
        "produces": ["application/xml"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "token"
          },
          {
            "in": "path",
            "name": "blogId",

            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Edit a blog post",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Blog"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "delete": {
        "tags": ["blog"],
        "summary": "Delete an existing blog",
        "description": "",
        "operationId": "deleteBlog",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "token"
          },
          {
            "in": "path",
            "name": "blogId",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/api/blogs/comment/{blogId}": {
      "post": {
        "tags": ["blog"],
        "summary": "Add a comment to an existing blog",
        "description": "",
        "operationId": "createComment",
        "consumes": ["application/json"],
        "produces": ["application/xml"],
        "parameters": [
          {
            "in": "path",
            "name": "blogId",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Adding a comment to a blog post",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Comment"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Blog not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
...
    "/api/user/profile/projects/{projectId}": {
      "delete": {
        "tags": ["profile"],
        "summary": "Find pet by ID",
        "description": "Delete a particular project",
        "operationId": "deleteProject",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "token"
          },
          {
            "in": "path",
            "name": "projectId",
            "description": "Delete a specified project",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Blog": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "content": {
          "type": "string"
        }
      }
    },
    "Comment": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "text": {
          "type": "string"
        }
      }
    },
...
  }
}